Feat: acceptation demande de transfert

+++ src/main/java/app/visa/repository/DemandeTransfertVisaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DemandeTransfertVisaRepository extends JpaRepository<DemandeTransfertVisa, Long> {
+public interface DemandeTransfertVisaRepository extends JpaRepository<DemandeTransfertVisa, Integer> {
 }
